pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "super-dev"
version = "2.3.0"
description = "Super Dev - Pipeline AI Coding Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Excellent", email = "11964948@qq.com"}
]
keywords = [
    "ai",
    "development",
    "agent",
    "workflow",
    "claude",
    "cli"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "ddgs>=9.0.0,<10.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "defusedxml>=0.7.1",
    "openai>=1.0.0,<2.0.0",
    "tenacity>=8.2.0",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn>=0.27.0",
    "tomli>=2.0.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/shangyankeji/super-dev"
Documentation = "https://github.com/shangyankeji/super-dev#readme"
Repository = "https://github.com/shangyankeji/super-dev"
Issues = "https://github.com/shangyankeji/super-dev/issues"

[project.scripts]
super-dev = "super_dev.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["super_dev*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.uv]
default-groups = ["dev"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:The 'app' shortcut is now deprecated\\.:DeprecationWarning:httpx\\._client",
]
